Experimentações bom backend e autenticação

+++ src/services/filme.service.ts
@@ -2,36 +2,74 @@ import { prisma } from "../prisma/client";
 
 export class FilmeService{
 
-    async create(dados: {titulo: string; nota: number; imagem: string; })
+    async create(dados: {titulo: string; nota: number; imagem: string;}, userId: number)
     {
         await prisma.filme.create({
-            data: dados,
+            data:{
+                ...dados,
+                userId,
+            },
         });
     }
 
-    async delete(id: number)
+    async delete(id: number, userId: number)
     {
+        const filme = await prisma.filme.findFirst({
+            where: {
+                id,
+                userId,
+            },
+        });
+
+        if (!filme) {
+            throw new Error(
+                "Filme não encontrado"
+            );
+        }
+
         await prisma.filme.delete({
-            where: {id}
+            where: {
+                id,
+            },
         });
     }
 
-    async findAll()
+    async findAll(userId: number)
     {
-        return await prisma.filme.findMany();
+        return await prisma.filme.findMany({
+            where: {
+                userId
+            }
+        });
     }
 
-    async findById(id: number)
+    async findById(id: number, userId: number)
     {
         return await prisma.filme.findUnique({
             where: {id}
         });
     }
 
-    async update(id:number, dados: {titulo: string; nota: number; imagem: string; })
+    async update(id:number, userId: number, dados: {titulo: string; nota: number; imagem: string; })
     {
+
+        const filme = await prisma.filme.findFirst({
+                where: {
+                    id,
+                    userId,
+                },
+            });
+
+            if (!filme) {
+                throw new Error(
+                    "Filme não encontrado"
+            );
+        }
+
         await prisma.filme.update({
-            where:{id},
+            where:{
+                id,
+            },
             data: dados,
         });
     }
